requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Social authentication
django-allauth==0.57.0

# Image handling
cloudinary==1.36.0
Pillow==10.0.1

# Database
psycopg2-binary==2.9.7  # For PostgreSQL (optional)

# Development
python-decouple==3.8
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
